requirements.txt
python-dotenv
PyMuPDF
requests
beautifulsoup4
selenium
webdriver-manager
pandas
openpyxl
supabase